Split generic function mutability tests
